\name{EnsDb.Hsapiens.v105}
\docType{package}

\alias{EnsDb.Hsapiens.v105-package}
\alias{EnsDb.Hsapiens.v105}
\alias{EnsDb.Hsapiens.v105}


\title{Ensembl based annotation package}

\description{
  This package loads an SQL connection to a database containing
  annotations from Ensembl. For examples and help on functions see the
  help pages from the \code{ensembldb} package!
}

\note{
  This data package was made from resources at Ensembl on
  Wed Feb  2 02:48:38 2022 and based on the 105
}

\author{Tao Wu}



\examples{
## load the library
##library(EnsDb.Hsapiens.v105)
## list the contents that are loaded into memory
ls('package:EnsDb.Hsapiens.v105')
## show the db object that is loaded by calling it's name
EnsDb.Hsapiens.v105

## for more examples see the ensembldb package.


}

\keyword{package}
\keyword{data}
